+++ src/quacc/calculators/espresso/__init__.py
@@ -1,0 +1,1 @@
+"""Enhanced version of the ASE Espresso calculator."""