+++ src/quacc/recipes/gaussian/core.py
@@ -9,6 +9,8 @@ from quacc import job
 from quacc.recipes.gaussian._base import base_fn
 
 if TYPE_CHECKING:
+    from pathlib import Path
+
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.cclib import cclibSchema
@@ -21,7 +23,7 @@ def static_job(
     spin_multiplicity: int = 1,
     xc: str = "wb97xd",
     basis: str = "def2tzvp",
-    copy_files: list[str] | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
     **calc_kwargs,
 ) -> cclibSchema:
     """
@@ -40,7 +42,7 @@ def static_job(
     basis
         Basis set
     copy_files
-        Files to copy to the runtime directory.
+        File(s) to copy to the runtime directory. If a directory is provided, it will be recursively unpacked.
     **calc_kwargs
         Custom kwargs for the Gaussian calculator. Set a value to
         `None` to remove a pre-existing key entirely. For a list of available
@@ -86,7 +88,7 @@ def relax_job(
     xc: str = "wb97xd",
     basis: str = "def2tzvp",
     freq: bool = False,
-    copy_files: list[str] | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
     **calc_kwargs,
 ) -> cclibSchema:
     """
@@ -107,7 +109,7 @@ def relax_job(
     freq
         If a frequency calculation should be carried out.
     copy_files
-        Files to copy to the runtime directory.
+        File(s) to copy to the runtime directory. If a directory is provided, it will be recursively unpacked.
     **calc_kwargs
         Custom kwargs for the Gaussian calculator. Set a value to
         `None` to remove a pre-existing key entirely. For a list of available