+++ src/quacc/recipes/psi4/core.py
@@ -14,6 +14,8 @@ except ImportError:
     psi4 = None
 
 if TYPE_CHECKING:
+    from pathlib import Path
+
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.ase import RunSchema
@@ -27,7 +29,7 @@ def static_job(
     spin_multiplicity: int = 1,
     method: str = "wb97x-v",
     basis: str = "def2-tzvp",
-    copy_files: list[str] | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
     **kwargs,
 ) -> RunSchema:
     """
@@ -46,7 +48,7 @@ def static_job(
     basis
         Basis set
     copy_files
-        Files to copy to the runtime directory.
+        File(s) to copy to the runtime directory. If a directory is provided, it will be recursively unpacked.
     **kwargs
         Custom kwargs for the Psi4 calculator. Set a value to
         `None` to remove a pre-existing key entirely. For a list of available