+++ tests/parsl/test_parsl_tutorials.py
@@ -183,25 +183,3 @@ def test_comparison3(tmp_path, monkeypatch):
     future2 = mult(future1, 3)
 
     assert future2.result() == 9
-
-
-def test_comparison4(tmp_path, monkeypatch):
-    monkeypatch.chdir(tmp_path)
-
-    @job
-    def add(a, b):
-        return a + b
-
-    @job
-    def make_more(val):
-        return [val] * 3
-
-    @subflow  #  (1)!
-    def add_distributed(vals, c):
-        return [add(val, c) for val in vals]
-
-    future1 = add(1, 2)
-    future2 = make_more(future1)
-    future3 = add_distributed(future2, 3)
-
-    assert future3.result() == [6, 6, 6]